Fix up AllTests
# Conflicts:
#	Tests/AppTests/AllTests.swift
