+++ Tests/AppTests/GitlabBuilderTests.swift
@@ -154,56 +154,3 @@ extension AllTests.GitlabBuilderTests {
     }
 
 }
-
-
-@Suite 
-
-
-struct LiveGitlabBuilderTests {
-
-    @Test(
-        .disabled("This is a live trigger test for end-to-end testing of pre-release builder versions")
-    )
-    func triggerBuild_live() async throws {
-        try await withDependencies {
-            // make sure environment variables are configured for live access
-            $0.environment.awsDocsBucket = { "spi-dev-docs" }
-            $0.environment.builderToken = {
-                // Set this to a valid value if you want to report build results back to the server
-                ProcessInfo.processInfo.environment["LIVE_BUILDER_TOKEN"]
-            }
-            $0.environment.buildTimeout = { 10 }
-            $0.environment.gitlabPipelineToken = {
-                // This Gitlab token is required in order to trigger the pipeline
-                ProcessInfo.processInfo.environment["LIVE_GITLAB_PIPELINE_TOKEN"]
-            }
-            $0.environment.siteURL = { "https://staging.swiftpackageindex.com" }
-            $0.httpClient = .liveValue
-        } operation: {
-            // set build branch to trigger on
-            Gitlab.Builder.branch = "main"
-
-            let buildId = UUID()
-
-            // use a valid uuid from a live db if reporting back should succeed
-            // SemanticVersion 0.3.2 on staging
-            let versionID = UUID(uuidString: "93d8c545-15c4-43c2-946f-1b625e2596f9")!
-
-            // MUT
-            let res = try await Gitlab.Builder.triggerBuild(
-                buildId: buildId,
-                cloneURL: "https://github.com/SwiftPackageIndex/SemanticVersion.git",
-                isDocBuild: false,
-                platform: .macosSpm,
-                reference: .tag(.init(0, 3, 2)),
-                swiftVersion: .v4,
-                versionID: versionID
-            )
-
-            print("status: \(res.status)")
-            print("buildId: \(buildId)")
-            print("webUrl: \(res.webUrl)")
-        }
-    }
-
-}
